ai-agent-cli/requirements.txt
typer[all]>=0.9.0
msal>=1.24.0
openai>=1.0.0
azure-identity>=1.14.0
httpx>=0.25.0
httpx-sse>=0.4.0
python-dotenv>=1.0.0
rich>=13.0.0

